food-ordering-system/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
bcrypt==4.1.2